Crate gl_utils

gl/glutin/glium + general graphics utility functions.

Re-exports

pub extern crate color_utils;

pub use self::render::Render;

pub use self::camera2d::Camera2d;

pub use self::camera3d::Camera3d;

Modules

camera2d
Types and functions for 2D rendering

camera3d
Types and functions for 3D rendering.

color
Color utilities.

graphics
General graphics-related utility functions.

info
Functions for writing various glium and glutin context info to a file handle.

mesh
Mesh generation utilities.

render
Support types for Glium rendering.

shader
Shader and shader program utilities.

texture
Texture related utilities.

vertex
Vertex related utilities including vertex specifications.

Macros

def_programs
Define a list of shaders with source paths and a list of shader programs using those shaders.

def_programs_include
Like def_programs! except that the shader source will be compiled-in with include_str!. The original source path is still available through the source_path method on shader IDs, but it is not required to be present when loading shaders.