Enum gl_utils::camera3d::Projection3dInner

pub enum Projection3dInner {
    Perspective {
        perspective_fov: PerspectiveFov<f32>,
        mat: Matrix4<f32>,
    },
    Orthographic {
        zoom: f32,
        ortho: Ortho<f32>,
        mat: Matrix4<f32>,
    },
}

Either a Perspective or Orthographic projection

Variants
Perspective
Fields of Perspective

perspective_fov: PerspectiveFov<f32>

Used to create the perspective projection matrix

mat: Matrix4<f32>

Constructed from the parameters in perspective_fov to transform points in view space to 4D homogenous clip coordinates based on a perspective projection

Orthographic
Fields of Orthographic

zoom: f32

Constant factor to zoom in and out orthographically.

The scale at zoom == 1.0 is defined by the constant ORTHOGRAPHIC_PIXEL_SCALE.

ortho: Ortho<f32>

Used to create the orthographic projection matrix

mat: Matrix4<f32>

Constructed from the parameters in ortho to transform points in view space to 4D homogenous clip coordinates based on an orthographic projection