Module gl_utils::info

Functions for writing various glium and glutin context info to a file handle.

glutin

The main datatypes in glutin are the glutin::GlWindow and glutin::HeadlessContext, both of which implement the glutin::GlContext trait:

glutin::GlWindow -- a glutin::Context and an associated glutin::Window which are constructed from glutin::ContextBuilders and glutin::WindowBuilders, respectively

glutin::HeadlessContext -- simply implements the glutin::GlContext trait; constructed from glutin::HeadlessRendererBuilders

glutin::Context -- an OpenGL context associated with a single Window or shared between multiple windows; implements glutin::GlContext and is constructed by glutin::ContextBuilder

glutin::Window -- represents a window with associated EventsLoop; can be built directly from a new glutin::EventsLoop or by passing it to glutin::WindowBuilder::build, or by passing the glutin::WindowBuilder itself (which configures a glutin::WindowAttributes struct) to glutin::GlWindow::new

glutin::EventsLoop -- retrieves events from the system and from the windows that were registered to the events loop

Traits:

glutin::GlContext -- represents an OpenGL context that can be shared between multiple windows; a new Context that is not yet shared can be only be acquired by passing a glutin::ContextBuilder (which configures a glutin::GlAttributes struct) to glutin::GlWindow::new

[Traits ]                         GlContext
                                     /\
                                     ||
[Structs]  GlAttributes       ==> HeadlessContext
           GlAttributes       ==> Context
           + WindowAttributes ==> GlWindow

GlContext --> PixelFormat
          --> Api
          --> is_current(), make_current()
          --> get_proc_address()
          --> resize()
          --> swap_buffers()

glutin::PixelFormat -- of a GlContext is the pixel format of the main framebuffer.

glutin::Api -- indicates whether OpenGl, OpenGlEs, or WebGl is being used

is_current, make_current -- queries and sets whether this context is the current one in this thread

get_proc_address -- required method for getting addresses of OpenGL functions used internally

resize -- required for contexts on some platforms (macOS, Wayland) to be manually updated when their window or surface is resized, i.e. it resize should be called for each Resized window event that is received

swap_buffers -- according to the documentation for glutin 0.12.0 (https://docs.rs/glutin/0.12.0/glutin/trait.GlContext.html#tymethod.swap_buffers), drivers may choose to override vsync settings to it is not possible to know in advance whether swap_buffers() will block or not.

glium

Glium extends the glutin::GlWindow and glutin::HeadlessContext datatypes with the glium::Display and glium::HeadlessRenderer datatypes, both of which dereference to a glium::backend::Context which is an implementor of glium::backend::Facade:

glium::Display -- (re-export of glium::backend::glutin::Display) a combination of a glutin::GlWindow and a glium::backend::Context

glium::HeadlessRenderer -- (re-export of glium::backend::glutin::headless::Headless) a glium::backend::Context without an associated window

Traits:

glium::backend::Facade -- a trait providing safe access for glium functions, implemented by glium::backend::Context and indirectly (via deref to a context) for glium::Display and glium::Headlessrenderer

[Traits ]                      Facade    CapabilitiesSource   Surface
                                 /\  \           /\             /\
                                 ||   \------\\\\||             ||
[Structs]     GlWindow     ==> Display  ====> Context =======> Frame
           HeadlessContext ==> Headless ====>  \\\\==========> DefaultFramebuffer
                                                \\\==========> EmptyFramebuffer,
                                                 \\==========> SimpleFramebuffer,
                                                  \==========> MultiOutputFramebuffer

Functions

glium_capabilities_source_info_write

glium_context_info_write

glium_display_info_write

glium_info_write

glium_surface_info_write

glutin_events_loop_info_write

glutin_gl_context_info_write

glutin_gl_window_info_write

glutin_info_write

glutin_window_info_write