Module gl_utils::mesh

Mesh generation utilities.

Functions

capsule_3d_instanced_lines_list
Two hemispherical "caps" connected at their equators.

capsule_3d_instanced_lines_list_counts

cylinder_3d_instanced_lines_list
Unit cylinder-- total height is 2.0 centered at 0.0 in the vertical (Z) axis, radius 1.0 in the X/Y plane.

cylinder_3d_instanced_lines_list_counts

grid_3d_instanced_lines_list
Produces vertices and indices for a 3D lines list arranged in a square grid in the X/Y plane of dims by dims dimensions.

grid_3d_instanced_lines_list_counts
Returns the (num_vertices, num_indices) for a mesh generated with grid_3d_instanced_lines_list

hemisphere_3d_instanced_lines_list
Generates a hemisphere with unit radius in the positive Z half-space with a number of parallels equal to latitude_divisions and a number of half-meridians equal to longitude_divisions

hemisphere_3d_instanced_lines_list_counts

sphere_3d_instanced_lines_list
Produces vertices and indices for a 3D lines list unit sphere (radius $1.0$) with the given number of latitudinal divisions rounded down to the nearest even number (ensuring the sphere always has an 'equator'), and with the given number of longitudinal divisions.

sphere_3d_instanced_lines_list_counts
Returns the (num_vertices, num_indices) for a mesh generated with sphere_3d_instanced_lines_list