Module gl_utils::render

Support types for Glium rendering.

The Render struct holds all the context information and resources required for making calls to OpenGL render commands (via glium).

Re-exports

pub use self::resource::Resource;

pub use self::resource::DefaultResource;

pub use self::viewport::Viewport;

Modules

params
Some useful drawing parameters.

resource
A trait for resource types with arbitrary draw functionality and a default implementation that supports switching to a "quad viewport" mode.

viewport
A rectangular viewport and associated 2D and 3D cameras.

Structs

Render
State for Glium-based rendering.

Functions

frame_fun_default
Default frame function clears all and calls resource draw_3d method followed by draw_2d method.

frame_fun_default_2d
Default 2D frame function clears all and calls resource draw_2d method.

frame_fun_default_3d
Default 3D frame function clears all and calls resource draw_3d method.