Module gl_utils::render::resource

A trait for resource types with arbitrary draw functionality and a default implementation that supports switching to a "quad viewport" mode.

Structs

DefaultResource
A default implementation of Resource containing the builtin shader programs and some pre-defined vertex sources.

ViewportTileData
Per-viewport data for tile rendering

Constants

CROSSHAIR_SPRITE_16X16_VERTEX_INDEX
Index into sprites_16x16_vertices for the crosshair.

LOWER_LEFT_VIEWPORT

LOWER_RIGHT_VIEWPORT

MAIN_VIEWPORT

MESH_3D_GRID_DIMS

TEXTURE_16X16_INDEX_CROSSHAIR
For use with BLEND_FUNC_NORMAL

TEXTURE_16X16_INDEX_CROSSHAIR_INVERSE
For use with BLEND_FUNC_INVERT_COLOR

TILESET_128X128_INDEX_EASCII_ACORN
For use with BLEND_FUNC_NORMAL

TILESET_128X128_INDEX_EASCII_ACORN_INVERSE
For use with BLEND_FUNC_INVERT_COLOR

UPPER_LEFT_VIEWPORT

UPPER_RIGHT_VIEWPORT

Traits

Resource
Represents a generic Glium resource type for a renderer to use as a source of drawing (vertex rendering).